+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # Evolución histórica de la informática
 
 ## v0.4.0-alpha1
-	- [2016-06-09] `App`: se agregó la estructura básica de la aplicación.
+- [2016-06-09]
+- [2016-06-09] `keyboard-events`: se agregó el polyfill para la API para `requestFullscreen`.
+- [2016-06-09] `App`: se agregó la estructura básica de la aplicación.